# apps/api/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: М3105 API
  version: "1.0"
  summary: Публичное API сайта группы М3105
  description: |
    Публичные эндпоинты сайта группы: предметы, лабораторные с дедлайнами, календарь и ICS-фид,
    конспекты, квизы, ЧаВо, статические страницы, поиск и расписание сдач.

    Все ответы в JSON (кроме `/calendar.ics` и файлов `/files/*`), время в RFC 3339 с таймзоной, даты без времени в формате `YYYY-MM-DD`.
    Возвращаются только опубликованные материалы. Авторизация для этих эндпоинтов не нужна.

    Эндпоинты админки (`/admin/*`) и аккаунтов студентов (`/auth/*`, `/me/*`, запись на сдачи)
    работают по HttpOnly-кукам и описаны в `docs/api.md` репозитория.
  contact:
    url: https://m3105.ru
security: []  # публичные эндпоинты без авторизации
servers:
  - url: https://m3105.ru/api/v1
    description: Продакшен
  - url: http://localhost:8080/api/v1
    description: Локальный API
...
paths:
...
  /settings:
    get:
      tags: [site]
      summary: Настройки сайта, текущая неделя и навигация
      operationId: getSettings
      responses:
        "200":
          description: Настройки
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SettingsResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
...
  /subjects:
    get:
      tags: [subjects]
      summary: Список предметов
      description: Предметы в порядке `position` со счётчиками опубликованных лаб, конспектов и квизов.
      operationId: listSubjects
      responses:
        "200":
          description: Предметы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SubjectWithCounts" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /subjects/{slug}:
    get:
      tags: [subjects]
      summary: Предмет с его лабами, конспектами и квизами
      operationId: getSubject
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Предмет
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubjectResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /labs:
    get:
      tags: [labs]
      summary: Список лабораторных
      description: Опубликованные лабы всех предметов или одного предмета. Длинные MDX-поля (`content`, `requirements`, `submission`, `variants`) в списке пустые.
      operationId: listLabs
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Лабы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Lab" }
        "4XX": { $ref: "#/components/responses/ClientError" }
...
  /notes:
    get:
      tags: [notes]
      summary: Список конспектов
      description: Опубликованные конспекты, по всем предметам или по одному. Поле `content` в списке пустое.
      operationId: listNotes
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Конспекты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Note" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /notes/{subject}/{slug}:
    get:
      tags: [notes]
      summary: Конспект целиком
      description: Вместе с привязанными квизами и соседними конспектами предмета для навигации.
      operationId: getNote
      parameters:
        - $ref: "#/components/parameters/SubjectSlug"
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Конспект
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /quizzes:
    get:
      tags: [quizzes]
      summary: Список квизов
      operationId: listQuizzes
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
      responses:
        "200":
          description: Квизы
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QuizSummary" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /quizzes/{slug}:
    get:
      tags: [quizzes]
      summary: Квиз с вопросами
      description: Формат вопросов описан в `docs/quiz-format.md`. Правильные ответы включены, проверка выполняется на клиенте.
      operationId: getQuiz
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Квиз
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "404": { $ref: "#/components/responses/NotFound" }
  /faq:
    get:
      tags: [content]
      summary: Частые вопросы
      operationId: listFAQ
      responses:
        "200":
          description: Вопросы в порядке `position`
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FAQItem" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /pages:
    get:
      tags: [content]
      summary: Статические страницы
      operationId: listPages
      responses:
        "200":
          description: Опубликованные страницы, `content` в списке пустой
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Page" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /pages/{slug}:
    get:
      tags: [content]
      summary: Статическая страница
      operationId: getPage
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Страница
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Page" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /practice:
    get:
      tags: [practice]
      summary: Занятия для сдачи лаб
      description: |
        По умолчанию только будущие. Для анонимных запросов `participants`, `myLabIds` и `availableLabs` пустые,
        `signedIn` равно `false`; вошедшие студенты видят участников и могут записаться (см. `docs/api.md`).
      operationId: listPractice
      parameters:
        - $ref: "#/components/parameters/SubjectFilter"
        - name: past
          in: query
          schema: { type: string, examples: ["1"] }
          description: Любое непустое значение включает прошедшие занятия
      responses:
        "200":
          description: Занятия
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PracticeListResponse" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /practice/{id}:
    get:
      tags: [practice]
      summary: Одно занятие
      operationId: getPractice
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Занятие
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PracticeSessionView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /comments/{target}/{id}:
    get:
      tags: [social]
      summary: Комментарии к конспекту, лабе или посту
      description: |
        Читать может кто угодно. `POST` с тем же путём и телом `{ body, attachmentIds? }` требует студенческую куку `edu_user`,
        `DELETE /comments/{id}` удаляет свой комментарий. Конспект или лаба должны быть опубликованы.
        `attachmentIds` — id своих файлов из `POST /files`, ещё ни к чему не прикреплённых; с файлами текст можно не писать.
      operationId: listComments
      parameters:
        - name: target
          in: path
          required: true
          schema: { type: string, enum: [note, lab, post] }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Комментарии, старые первыми
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }
        "404": { $ref: "#/components/responses/NotFound" }
  /posts:
    get:
      tags: [social]
      summary: Лента постов
      description: |
        Пользовательские посты одной ленты. `POST /posts` с `PostInput` создаёт пост от имени студента,
        `PUT`/`DELETE /posts/{id}` меняют и удаляют свой пост, `PUT`/`DELETE /posts/{id}/like` ставят и снимают лайк.
        Поля `liked` и `mine` считаются для студента, если передана кука `edu_user`.
        Без куки отдаются только посты с `visibility: public`: `members` видят лишь вошедшие студенты,
        и посты с `nsfw: true` всегда `members`. Комментарии к скрытому посту анонимам тоже недоступны (404),
        как и файлы такого поста.
      operationId: listPosts
      parameters:
        - name: kind
          in: query
          required: true
          schema: { type: string, enum: [shawarma, joke] }
        - name: sort
          in: query
          schema: { type: string, enum: [new, top], default: new }
          description: "`top` — сначала посты с наибольшим числом лайков"
      responses:
        "200":
          description: Посты
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Post" }
        "4XX": { $ref: "#/components/responses/ClientError" }
  /files/{id}/{name}:
    get:
      tags: [files]
      summary: Файл вложения
      description: |
        Адрес файла приходит в поле `url` у `Attachment`; последний сегмент — транслит имени, сервер смотрит только на `id`.
        Картинки, PDF, видео и аудио открываются в браузере, остальное скачивается; `?download=1` скачивает всегда.
        `?w=160|320|640|1280|1920` у JPEG, PNG и WebP отдаёт уменьшенную копию (WebP, если он есть в `Accept`, `Vary: Accept`);
        картинка уже нужной ширины, GIF, SVG и остальные типы приходят оригиналом.
        Поддерживаются `Range` и `ETag`. Файлы поста «для своих» анонимам не отдаются (404),
        как и загрузки, ещё не прикреплённые к комментарию или посту, — их видит только автор.

        Загрузка: `POST /files` (студент, кука `edu_user`, картинки JPG/PNG/GIF/WebP и PDF до 10 МБ, 30 файлов за 12 часов)
        и `POST /admin/files` (админка, любые файлы до 50 МБ), тело `multipart/form-data` с полем `file`, ответ — `Attachment`.
        У фото вырезаются EXIF и прочие метаданные, ориентация сохраняется.
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, pattern: "^[a-z2-7]{26}$" }
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: download
          in: query
          schema: { type: string, enum: ["1"] }
          description: Отдать как вложение, а не открыть в браузере
        - name: w
          in: query
          schema: { type: integer, enum: [160, 320, 640, 1280, 1920] }
          description: Ширина превью картинки; другие значения игнорируются
      responses:
        "200":
          description: Содержимое файла
          headers:
            Content-Disposition:
              schema: { type: string, examples: ["inline; filename=photo.jpg"] }
            Cache-Control:
              schema: { type: string, examples: ["public, max-age=31536000, immutable"] }
          content:
            "*/*":
              schema: { type: string, format: binary }
        "206":
          description: Часть файла по заголовку `Range`
        "404": { $ref: "#/components/responses/NotFound" }
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9-]+$" }
      description: Слаг, сгенерированный из названия транслитерацией
    SubjectSlug:
      name: subject
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9-]+$" }
      description: Слаг предмета
    SubjectFilter:
      name: subject
      in: query
      schema: { type: string }
      description: Слаг предмета для фильтрации
  responses:
    ClientError:
      description: Некорректный запрос или превышен лимит; формат ошибки одинаковый для всех кодов
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Не найдено или не опубликовано
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Не найдено", code: not_found }
  schemas:
    Error:
      type: object
      description: Формат всех ошибок API
      properties:
        error: { type: string, description: "Сообщение для пользователя" }
        code:
          type: string
          enum: [validation, not_found, conflict, unauthorized, bad_password, rate_limited, internal]
        fields:
          type: object
          additionalProperties: { type: string }
          description: Ошибки валидации по полям (только для `validation`)
      required: [error, code]
    DateTime:
      type: string
      format: date-time
      description: RFC 3339 с таймзоной
    Date:
      type: string
      format: date
      description: "`YYYY-MM-DD`"
    Parity:
      type: string
      enum: [odd, even]
    Status:
      type: string
      enum: [draft, published]
      description: В публичных ответах всегда `published`
...
    Link:
      type: object
      properties:
        title: { type: string }
        url: { type: string, format: uri }
      required: [title, url]
    Settings:
      type: object
      properties:
        siteTitle: { type: string }
        groupName: { type: string }
        description: { type: string }
        semesterStart: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
        semesterEnd: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
        firstWeekParity: { $ref: "#/components/schemas/Parity" }
        timezone: { type: string, examples: [Europe/Moscow] }
        links:
          type: array
          items: { $ref: "#/components/schemas/Link" }
        inviteCode: { type: string, description: "В публичных ответах всегда пустая строка" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
      required: [siteTitle, groupName, description, semesterStart, semesterEnd, firstWeekParity, timezone, links, inviteCode, updatedAt]
    Week:
      type: object
      description: Текущая учебная неделя, считается от начала семестра
      properties:
        number: { type: integer, description: "Номер недели; 0, если семестр не настроен" }
        parity: { $ref: "#/components/schemas/Parity" }
        start: { $ref: "#/components/schemas/DateTime" }
        inRange: { type: boolean, description: "Сейчас идёт семестр" }
        configured: { type: boolean, description: "Дата начала семестра задана" }
      required: [number, parity, start, inRange, configured]
    AuthInfo:
      type: object
      properties:
        telegramBot: { type: string, description: "Username бота; пусто, если вход через Telegram не настроен" }
        telegramBotId: { type: string }
        devLogin: { type: boolean, description: "Доступен вход по имени (только вне production)" }
        inviteRequired: { type: boolean, description: "Для первого входа нужен код доступа" }
      required: [telegramBot, telegramBotId, devLogin, inviteRequired]
...
    Subject:
      allOf:
        - $ref: "#/components/schemas/SubjectRef"
        - type: object
          properties:
            teacher: { type: string }
            description: { type: string, description: "MDX" }
            links:
              type: array
              items: { $ref: "#/components/schemas/Link" }
            position: { type: integer }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
          required: [teacher, description, links, position, createdAt, updatedAt]
    SubjectWithCounts:
      allOf:
        - $ref: "#/components/schemas/Subject"
        - type: object
          properties:
            labsCount: { type: integer }
            notesCount: { type: integer }
            quizzesCount: { type: integer }
          required: [labsCount, notesCount, quizzesCount]
...
    Lab:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { type: integer }
            number: { type: integer, minimum: 0, description: "Ноль допустим для необязательной вводной лабы" }
            slug: { type: string }
            title: { type: string }
            summary: { type: string }
            content: { type: string, description: "Задание, MDX" }
            requirements: { type: string, description: "Требования, MDX" }
            submission: { type: string, description: "Как сдавать, MDX" }
            variants: { type: string, description: "Варианты, MDX" }
            materials:
              type: array
              items: { $ref: "#/components/schemas/Link" }
            deadlineAt: { oneOf: [{ $ref: "#/components/schemas/DateTime" }, { type: "null" }] }
            deadlineNote: { type: string }
            maxScore: { oneOf: [{ type: integer }, { type: "null" }] }
            teacher: { type: string }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
          required: [id, subjectId, number, slug, title, summary, content, requirements, submission, variants, materials, deadlineAt, deadlineNote, maxScore, teacher, status, createdAt, updatedAt]
    Note:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { type: integer }
            number: { type: integer, minimum: 1 }
            slug: { type: string }
            title: { type: string }
            summary: { type: string }
            content: { type: string, description: "MDX с формулами и кодом" }
            lectureDate: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
            quizzesCount: { type: integer, description: "Опубликованные квизы, привязанные к конспекту" }
          required: [id, subjectId, number, slug, title, summary, content, lectureDate, status, createdAt, updatedAt, quizzesCount]
    NoteResponse:
      allOf:
        - $ref: "#/components/schemas/Note"
        - type: object
          properties:
            quizzes:
              type: array
              items: { $ref: "#/components/schemas/QuizSummary" }
            prev: { oneOf: [{ $ref: "#/components/schemas/Note" }, { type: "null" }] }
            next: { oneOf: [{ $ref: "#/components/schemas/Note" }, { type: "null" }] }
          required: [quizzes, prev, next]
...
    Quiz:
      allOf:
        - $ref: "#/components/schemas/SubjectJoin"
        - type: object
          properties:
            id: { type: integer }
            subjectId: { oneOf: [{ type: integer }, { type: "null" }] }
            noteId: { oneOf: [{ type: integer }, { type: "null" }] }
            slug: { type: string }
            title: { type: string }
            description: { type: string }
            questions:
              type: array
              items: { $ref: "#/components/schemas/QuizQuestion" }
            shuffleQuestions: { type: boolean }
            shuffleOptions: { type: boolean }
            status: { $ref: "#/components/schemas/Status" }
            createdAt: { $ref: "#/components/schemas/DateTime" }
            updatedAt: { $ref: "#/components/schemas/DateTime" }
            noteSlug: { type: string }
            noteTitle: { type: string }
          required: [id, subjectId, noteId, slug, title, description, questions, shuffleQuestions, shuffleOptions, status, createdAt, updatedAt, noteSlug, noteTitle]
    QuizSummary:
      allOf:
        - $ref: "#/components/schemas/Quiz"
        - type: object
          properties:
            questionsCount: { type: integer }
          required: [questionsCount]
    FAQItem:
      type: object
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string, description: "MDX" }
        category: { type: string }
        position: { type: integer }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
      required: [id, question, answer, category, position, createdAt, updatedAt]
    Page:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        summary: { type: string }
        content: { type: string, description: "MDX" }
        position: { type: integer }
        showInNav: { type: boolean, description: "Показывать в меню сайта" }
        status: { $ref: "#/components/schemas/Status" }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
      required: [id, slug, title, summary, content, position, showInNav, status, createdAt, updatedAt]
...
    SettingsResponse:
      type: object
      properties:
        settings: { $ref: "#/components/schemas/Settings" }
        week: { $ref: "#/components/schemas/Week" }
        now: { $ref: "#/components/schemas/DateTime" }
        navPages:
          type: array
          items: { $ref: "#/components/schemas/Page" }
          description: Страницы с `showInNav`, без `content`
        auth: { $ref: "#/components/schemas/AuthInfo" }
      required: [settings, week, now, navPages, auth]
...
    SubjectResponse:
      type: object
      properties:
        subject: { $ref: "#/components/schemas/Subject" }
        labs:
          type: array
          items: { $ref: "#/components/schemas/Lab" }
        notes:
          type: array
          items: { $ref: "#/components/schemas/Note" }
        quizzes:
          type: array
          items: { $ref: "#/components/schemas/QuizSummary" }
      required: [subject, labs, notes, quizzes]
...
    PracticeSessionView:
      allOf:
        - $ref: "#/components/schemas/PracticeSession"
        - type: object
          properties:
            participants:
              type: array
              items: { $ref: "#/components/schemas/Participant" }
              description: Только для вошедших студентов, иначе пусто
            myLabIds:
              type: array
              items: { type: integer }
            full: { type: boolean, description: "Мест нет" }
            past: { type: boolean, description: "Занятие уже прошло" }
            availableLabs:
              type: array
              items: { $ref: "#/components/schemas/LabRef" }
              description: Опубликованные лабы предмета, только для вошедших
          required: [participants, myLabIds, full, past, availableLabs]
    PracticeListResponse:
      type: object
      properties:
        sessions:
          type: array
          items: { $ref: "#/components/schemas/PracticeSessionView" }
        now: { $ref: "#/components/schemas/DateTime" }
        signedIn: { type: boolean }
      required: [sessions, now, signedIn]
    Comment:
      type: object
      properties:
        id: { type: integer }
        targetType: { type: string, enum: [note, lab, post] }
        targetId: { type: integer }
        body: { type: string, maxLength: 2000 }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
        authorId: { type: integer }
        authorName: { type: string }
        authorPhotoUrl: { type: string }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        mine: { type: boolean, description: Комментарий текущего студента }
      required: [id, targetType, targetId, body, createdAt, updatedAt, authorId, authorName, authorPhotoUrl, attachments, mine]
    Post:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string, enum: [shawarma, joke] }
        title: { type: string, description: Название точки для шавермы, необязательный заголовок для анекдота }
        body: { type: string, maxLength: 4000 }
        address: { type: string }
        price: { type: [integer, "null"], description: Цена в рублях }
        rating: { type: [integer, "null"], minimum: 1, maximum: 5 }
        visibility: { type: string, enum: [public, members], description: "`members` — только для вошедших студентов" }
        nsfw: { type: boolean, description: Пометка 18+; сайт размывает текст до подтверждения возраста }
        createdAt: { $ref: "#/components/schemas/DateTime" }
        updatedAt: { $ref: "#/components/schemas/DateTime" }
        authorId: { type: integer }
        authorName: { type: string }
        authorPhotoUrl: { type: string }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        liked: { type: boolean }
        mine: { type: boolean }
      required: [id, kind, title, body, address, price, rating, visibility, nsfw, createdAt, updatedAt, authorId, authorName, authorPhotoUrl, likesCount, commentsCount, attachments, liked, mine]
...